Clean-up and adding go.mod + go.sum
